% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{speech_char_count}
\alias{speech_char_count}
\title{Character count after removing non-speech sections}
\usage{
speech_char_count(textstring)
}
\arguments{
\item{textstring}{character vector}
}
\value{
integer, number of characters in the speech
}
\description{
Character count after removing non-speech sections
}
