% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{speech_clean}
\alias{speech_clean}
\title{Remove non-speech sections from speech}
\usage{
speech_clean(textstring)
}
\arguments{
\item{textstring}{character vector}
}
\value{
cleaned speech
}
\description{
To be written
}
