% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_capacity}
\alias{extract_capacity}
\title{Remove non-speech sections from speech}
\usage{
extract_capacity(textstring)
}
\arguments{
\item{textstring}{character vector}
}
\value{
cleaned speech
}
\description{
To be written
}
