% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_capacity}
\alias{extract_capacity}
\title{Extract speaker capacity from speech text.}
\usage{
extract_capacity(textstring, other = TRUE, plenary = FALSE,
  hoc = FALSE)
}
\arguments{
\item{textstring}{speech text}

\item{other}{bool. whether unmached text will be replaced to "Other"}

\item{plenary}{bool. whether the speech is from plenary session}

\item{hoc}{bool. the speech text is from Upper House. default \code{FALSE}}
}
\value{
text of speaker capacity
}
\description{
To be written
}
